pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Webpage SEO analyzer"
authors = ["Pavel Krivosheev <pkrivosheev@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.28.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
